Fix of the issue #2922: now webcams can be added to geopaths
Additionally there is a new config entry with list of types forbidden in geopaths and small code refactors.
